requirements.txt
streamlit
pandas
Pillow
requests